requirements.txt
abseil-cpp==20210324.1
arrow-cpp==4.0.0
asn1crypto==1.4.0
aws-c-cal==0.5.11
aws-c-common==0.6.2
aws-c-event-stream==0.2.7
aws-c-io==0.10.5
aws-checksums==0.1.11
aws-sdk-cpp==1.8.186
azure-common==1.1.27
azure-core==1.14.0
azure-storage-blob==12.8.1
beautifulsoup4==4.9.3
blinker==1.4
bodo==2021.7.1
boto3==1.18.19
botocore==1.21.19
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
c-ares==1.17.2
ca-certificates==2021.5.30
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.0
cryptography==3.4.7
cycler==0.10.0
filprofiler==2021.05.0
freetype==2.10.4
fsspec==2021.7.0
gflags==2.2.2
glog==0.5.0
gmp==6.2.1
greenlet==1.1.1
grpc-cpp==1.38.0
hdf5==1.10.6
idna==2.10
isodate==0.6.0
jbig==2.1
jmespath==0.10.0
jpeg==9d
kiwisolver==1.3.1
krb5==1.19.2
lcms2==2.12
lerc==2.2.1
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libcurl==7.78.0
libcxx==12.0.1
libdeflate==1.7
libedit==3.1.20191231
libev==4.33
libevent==2.1.10
libffi==3.3
libgfortran==5.0.0
libgfortran5==9.3.0
liblapack==3.9.0
libllvm10==10.0.1
libnghttp2==1.43.0
libopenblas==0.3.17
libpng==1.6.37
libprotobuf==3.16.0
libssh2==1.9.0
libthrift==0.14.1
libtiff==4.3.0
libutf8proc==2.6.1
libwebp-base==1.2.1
llvm-openmp==12.0.1
llvmlite==0.36.0
lz4-c==1.9.3
matplotlib==3.4.3
matplotlib-base==3.4.3
memory_profiler==0.58.0
mpi==1.0
mpi4py==3.0.3
mpich==3.3.2
msrest==0.6.21
ncurses==6.2
numba==0.53.0
numpy==1.21.1
oauthlib==3.1.1
olefile==0.46
openjpeg==2.4.0
openssl==1.1.1k
orc==1.6.8
oscrypto==1.2.1
pandas==1.2.5
parquet-cpp==1.5.1
pillow==8.3.1
pip==21.2.2
psutil==5.8.0
pyarrow==4.0.0
pycparser==2.20
pycryptodomex==3.10.1
pyjwt==2.1.0
pyopenssl==20.0.1
pyparsing==2.4.7
pysocks==1.7.1
python==3.9.6
python-dateutil==2.8.2
python_abi==3.9
pytz==2020.5
re2==2021.04.01
readline==8.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
setuptools==52.0.0
six==1.16.0
snappy==1.1.8
snowflake-connector-python==2.4.4
snowflake-sqlalchemy==1.3.1
soupsieve==2.0.1
sqlalchemy==1.4.22
sqlite==3.36.0
threadpoolctl==2.2.0
tk==8.6.10
tornado==6.1
tzdata==2021a
urllib3==1.26.6
wheel==0.36.2
xz==5.2.5
yappi==1.3.2
zlib==1.2.11
zstd==1.5.0